count aggregator should return Int64 not Integer

+++ Opaleye/Aggregate.hs
@@ -9,6 +9,8 @@ import qualified Opaleye.Column as C
 
 import qualified Database.HaskellDB.PrimQuery as HPQ
 
+import           GHC.Int (Int64)
+
 -- | Group the aggregation by equality on the input to 'groupBy'.
 groupBy :: Aggregator (C.Column a) (C.Column a)
 groupBy = A.makeAggr' Nothing
@@ -17,13 +19,8 @@ groupBy = A.makeAggr' Nothing
 sum :: Aggregator (C.Column a) (C.Column a)
 sum = A.makeAggr HPQ.AggrSum
 
--- TODO: We have to decide what is the most appropriate return type
--- for the count aggregator.  In Postgres it returns a 64 bit integer,
--- but making this explicit will imply a lot of annoying fiddling on
--- the part of our users.  Can we get away with just saying 'Integer'
--- here?
 -- | Count the number of non-null rows in a group.
-count :: Aggregator (C.Column a) (C.Column Integer)
+count :: Aggregator (C.Column a) (C.Column Int64)
 count = A.makeAggr HPQ.AggrCount
 
 {-|
